requirements/common.txt
Click==8.1.3
pydantic==1.10.9
httpx==0.24.1
websockets==11.0.3
auth0-python==4.3.0
typer==0.9.0
PyJWT==2.7.0
requests==2.31.0
shellingham==1.5.0
rich==13.4.2
urllib3==1.26.6
colorama==0.4.4
wheel==0.36.2
jsonlines==3.1.0
requests_oauthlib==1.3.1
platformdirs==3.9.1
